Move download logic into his own component, clean AppBrowser.js
